+++ sideline/app/tendencies/loading.tsx
@@ -1,5 +1,5 @@
-import { FilmRoomSkeleton } from "@/components/shared/PageSkeleton";
+import { TendenciesHomeSkeleton } from "@/components/shared/PageSkeleton";
 
 export default function TendenciesLoading() {
-  return <FilmRoomSkeleton />;
+  return <TendenciesHomeSkeleton />;
 }